fix windows running process detection

+++ chat/chat.go
@@ -371,21 +371,7 @@ func (b *Bounce) ensureOnlyOneInstance() {
 
 	process, err := os.FindProcess(pid)
 	if err != nil {
-		// Always true on unix systems, on windows an error means the process is not running
-		// https://stackoverflow.com/a/15204759
-		// TODO: delete the pid file and return here if on windows?
-		log.WithFields(log.Fields{
-			"pid":   pid,
-			"error": err.Error(),
-		}).Warn("error on os.FindProcess")
-	} else {
-		err := process.Signal(syscall.Signal(0))
-		if err == nil {
-			log.WithFields(log.Fields{
-				"pid": pid,
-			}).Fatal("Another instance of Bounce is running.  Please close it, or if you are sure it is not running, delete the pid file and try again: ", pidFile)
-		} else if err.Error() == "no such process" || err.Error() == "os: process already finished" {
-			// Delete the old pid file that refers to a dead process
+		if runtime.GOOS == "windows" {
 			err = os.Remove(pidFile)
 			if err != nil {
 				log.WithFields(log.Fields{
@@ -393,19 +379,46 @@ func (b *Bounce) ensureOnlyOneInstance() {
 					"error": err.Error(),
 				}).Fatal("error deleting old pid file")
 			}
-			// Write the current PID to file
-			err = os.WriteFile(pidFile, []byte(strconv.Itoa(os.Getpid())), 0600) // TODO: sync/flush?
-			if err != nil {
-				log.WithFields(log.Fields{
-					"error": err.Error(),
-					"path":  pidFile,
-				}).Fatal("error writing current pid file")
-			}
 		} else {
 			log.WithFields(log.Fields{
-				"pid":           pid,
-				"syscall_error": err.Error(),
+				"pid":   pid,
+				"error": err.Error(),
+			}).Error("os.FindProcess returned an error on a non-windows system")
+		}
+	} else {
+		if runtime.GOOS == "windows" {
+			log.WithFields(log.Fields{
+				"pid": pid,
 			}).Fatal("Another instance of Bounce is running.  Please close it, or if you are sure it is not running, delete the pid file and try again: ", pidFile)
+		} else {
+			err := process.Signal(syscall.Signal(0))
+			if err == nil {
+				log.WithFields(log.Fields{
+					"pid": pid,
+				}).Fatal("Another instance of Bounce is running.  Please close it, or if you are sure it is not running, delete the pid file and try again: ", pidFile)
+			} else if err.Error() == "no such process" || err.Error() == "os: process already finished" {
+				// Delete the old pid file that refers to a dead process
+				err = os.Remove(pidFile)
+				if err != nil {
+					log.WithFields(log.Fields{
+						"path":  pidFile,
+						"error": err.Error(),
+					}).Fatal("error deleting old pid file")
+				}
+				// Write the current PID to file
+				err = os.WriteFile(pidFile, []byte(strconv.Itoa(os.Getpid())), 0600) // TODO: sync/flush?
+				if err != nil {
+					log.WithFields(log.Fields{
+						"error": err.Error(),
+						"path":  pidFile,
+					}).Fatal("error writing current pid file")
+				}
+			} else {
+				log.WithFields(log.Fields{
+					"pid":           pid,
+					"syscall_error": err.Error(),
+				}).Fatal("Another instance of Bounce may be running.  Please close it, or if you are sure it is not running, delete the pid file and try again: ", pidFile)
+			}
 		}
 	}
 }
